[ccore] Correction for CCORE.

+++ ccore/src/interface/pyclustering_package.cpp
@@ -32,7 +32,7 @@ pyclustering_package::pyclustering_package(void) :
 
 pyclustering_package::pyclustering_package(const pyclustering_type_data package_type) :
     size(0),
-    type(package_type),
+    type((unsigned int) package_type),
     data(nullptr)
 { }
